Cast max_trip_distance to string in batch job params
Submitting a job on staging wasn't working, which turned out to be because the
boto3 client `submit_job` call was failing with:
  File "/usr/local/lib/python3.6/site-packages/botocore/client.py", line 357, in _api_call
    return self._make_api_call(operation_name, kwargs)
  File "/usr/local/lib/python3.6/site-packages/botocore/client.py", line 634, in _make_api_call
    api_params, operation_model, context=request_context)
  File "/usr/local/lib/python3.6/site-packages/botocore/client.py", line 682, in _convert_to_request_dict
    api_params, operation_model)
  File "/usr/local/lib/python3.6/site-packages/botocore/validate.py", line 297, in serialize_to_request
    raise ParamValidationError(report=report.generate_report())
botocore.exceptions.ParamValidationError: Parameter validation failed:
Invalid type for parameter containerOverrides.environment[17].value, value: 2680, type: <class 'int'>, valid types: <class 'str'>

That's NB_MAX_TRIP_DISTANCE, which is being passed as an integer, but boto
wants all those params to be strings, apparently. So this casts it to
a string.