Send undefined when max trip distance is null
I noticed while testing non-US cities that setting a max trip distance then
clearing it didn't make it revert to the default, it caused it to be set to
None, which led to a crash.  This catches that case and makes sure that a
cleared field will send the same thing as a field that was never modified.