Cast values in batch environment to strings
We pass a bunch of parameters to the Batch job as environment variables,
and the values need to be strings or the API call will fail.
This has caught us before, and it just caught us again, so rather than
have to remember to cast each value when we add or change them, this
adjusts the function that assembles the parameter for the Boto call so
that it will cast every value to a string.  Calling `str` on something
that's already a string won't do any harm, so there's no reason not to
do it just in case.