Save zero rather than null for tf_seg_str in NeighborhoodWaysResults
Apparently the Django upgrade changed the behavior of [LayerMapping](https://docs.djangoproject.com/en/3.2/ref/contrib/gis/layermapping/)
so that it saves null values from the shapefile for nullable integer fields
as nulls in the database rather than zeroes. Which makes sense, but our
tile config is counting on those zeroes.

Rather than change the tiler config to handle nulls in addition to zeroes,
this adjusts the command that reimports the stress shapefile to just change
all nulls to zeroes in the newly-loaded data.