Port from Jenkins to Github Actions
Most items copied over with only minor changes, except:

Commenting out update_batch_definitions as that would need a good number of changes
to support not using an AWS profile in Github and we are looking into improving that
in the future anyway.

Slack notifications for build success or failure are covered when pushing a PR
by the Github integration.

Resolves: #882