Fix cleanup commands for non-US population files
When downloading a population file from S3, the file gets copied to the
temporary directory and needs to be cleaned up, but when using the default
local file (/data/population.zip) it doesn't, so the `rm` command to remove
it fails.  Also, the command to rename the unzipped files to the default
filename is necessary if they're not already named that, but crashes if they
are. This changes both commands so they'll still do their thing but will be
more tolerant of the files being missing or already named correctly.