Add field to set neighborhood speed limit directly
Adds a field to the Neighborhood model to explicitly set a speed limit.
It will be provided to the analysis and stored in the speed limit output
file, the same way we do for speed limits looked up by FIPS code.

This is primarily intended for non-US cities, where there aren't state and
city FIPS codes to set speed limits by. It's also not supported by the batch
job process--it can only be set on individual neighborhoods--but since non-US
cities aren't supported by that process for other reasons, that should be fine.

The FIPS lookups will still happen, but if a speed limit has been manually set,
it will be used for the CITY_SPEED value, regardless of whether the
neighborhood has a city FIPS code with an associated default limit. The form
is set up to only allow one or the other, but in case that's not foolproof or
something changes in the future, it seems sensible to have the one that's
specified directly take precedence.

The limit should be in MPH, even if the actual local speed limit is in km/h.
The km/h value will be calculated, rounded to the nearest 5, for display.