Add '--rm' flag to run-local-analysis command
We didn't include the --rm flag in the `docker-compose run` command in
this script because it seemed worthwhile to keep containers around in case
we wanted to examine them further after a test run. But in fact, we almost
never do that--the `bash` at the end gives us a chance to examine them
when they've finished but before final exit, and once we've closed that
out it's probably because we're planning to change something and do a new
test. So having the containers stick around is doing more harm (filling up
the disk and requiring cleanup) than good. This adds the flag so they'll
get cleaned up instead.