Stop trying to cache non-US population files
The cache setup for population files relies on the fact that they have their
FIPS code in the name so it's not really possible to get them mixed up.
That doesn't hold for non-US neighborhoods, where we just call them all
"population". So we shouldn't have been caching them, but we were, resulting
in jobs failing because they ended up with the population file for the wrong
country.

This doesn't really make that whole setup less complicated, but hopefully it
makes it more correct. It will now only check cache and upload files to cache
for US jobs.

This also makes it so that `POP_FILE_URL` won't get ignored for US jobs.
I'm not sure we need it, but since the form field is always there, using
it rather than silently ignoring it if it's filled in seems better.

Resolves #910