Downgrade django-countries to fix country search
We use a package, `django-countries`, to provide the country field for
the Neighborhood model, and it turns out they changed how `icontains`
searching against that field works (see changelog:
https://github.com/SmileyChris/django-countries/blob/master/CHANGES.rst#73-28-february-2022)
so that it compares against the full name of the country. The field
stores the alpha-2 abbreviation, and until the change, `icontains`
queries just used that value.  In the new version, the they use the full
name of the country and *don't* match an alpha-2 code (unless it happens
to be a substring of the full name).

We could switch to using `iexact` instead, but that has the disadvantage that
typing just one letter will clear the list, rather than reducing it to
countries that contain that letter in their abbreviation. Typing the second
letter gets you the list you're interested in, but the brief moment where
the list is empty feels weird.

So this fixes the issue by downgrading `django-countries` to the last
version with the old behavior.

Resolves #908.