Adjust score weights when there's no jobs data (#914)
When we skip calculating employment accessibility because there's no jobs
data available, we also need to adjust the combined scores that the employment
score feeds into. So this changes the opportunity score, which is a
weighted average of employment and education-related scores, and the overall
score to avoid averaging in a zero. The former also needs divide-by-zero
protection, since with this change it becomes possible for none of the
components that go into the opportunity score to be present.