Only get state OSM extracts for US states
The OSM downloader was treating any state abbreviation as a US state and
trying to download a state extract for it.  But now there are non-US
neighborhoods with states that are using this script, so it needed to have
that logic protected by a "only in the US" check.

Geofabrik has state-level extracts for several other countries, but knowing
which countries it has and which it doesn't, and getting the names to match
up exactly right, would be fiddly and error-prone, so we'll just stick with
national extracts.

Resolves #915