Store crashes in 'geography' mode and add neighborhood geography for comparison
Since we want to do a DWithin lookup on crashes without having to deal with local
projections, storing their locations in a geography field makes sense. To make
the query actually fast, we also need an indexed geography field for the
neighborhood boundaries we're going to compare them against.

Notes:
- I think it would be fine to just convert Neighborhood.geom to a geography field,
but I'm not sure, and adding a copy seems fine, too.
- This changes migration 48 rather than adding a new migration to convert the
Crash field to geography. That's because reversing a geom->geog migration causes
a crash. It says
  > django.db.utils.ProgrammingError: column "geom_pt" cannot be cast automatically to type geometry
  > HINT:  You might need to specify "USING geom_pt::geometry(Point,4326)".
but there's no way that I can tell to override the `AlterField` reversal to
actually do that, so just changing the original field seems like the next way
to go.