Increase analysis task RAM to 61 GiB and increase PG settings
We've been running this job with 8 vCPUs but only 31 GiB of RAM, which doesn't
make a ton of sense because that's all the CPU but only half the RAM of an
i3.2xlarge instance. To be fair, we didn't think we would actually get close to
filling up even the 31 GiB, but apparently we are on the largest jobs, so it's
time to allocate the rest.

I also:
- Increased the parameters we apply to the PostgreSQL instance within the
  analysis container, to make more use of the copious resources we're providing
  to these tasks. Some of the values are probably overkill, but there's no reason
  not to use the resources, and some could make the jobs faster or more robust.
- Matched all the parameters in the staging job definiton to the ones in
  production. We seldom run staging jobs, so making them smaller doesn't appreciably
  help with keeping costs down, and it seems valuable to make the test environment
  as similar to production as we practically can.