Increase analysis task RAM to 61 GiB and match staging to prod
We've been running this job with 8 vCPUs but only 31 GiB of RAM, which doesn't
make a ton of sense because that's all the CPU but only half the RAM of an i3.2xl
instance. To be fair, we didn't think we would actually get close to filling up
even the 31 GiB, but apparently we are, so it's time to allocate the rest.

I also matched all the parameters in the staging job definiton to the ones in
production. We seldom run staging jobs, so making them smaller doesn't appreciably
help with keeping costs down, and it seems valuable to make the test environment
as similar to production as we practically can.