Strip backslash characters from OSM segment names
It turns out osm2pgrouting *really* doesn't like it if you have a backslash in
a segment name.  It produces an error that knocks out not just the segment in
question, but everything that the tool would have otherwise loaded in the chunk
that it's processing.
Here's a Stack Overflow about it: https://gis.stackexchange.com/a/12873
but the recommendation there is to modify the osm2pgrouting code. I just added
a command to the import_osm.sh script to use 'sed' to strip out any backslash
characters.