Add fixture with sample analysis results
Adds some lines to the 'update' script to download a data fixture containing
a few neighborhoods and analysis results and import it into the database.

Because the import is a bit slow and only needs to be done on initial setup,
this also modifies the 'update' script to take a '--load-data' argument, so
that it will only load the fixture (and load crash data, which also takes
some time and doesn't need to be repeated) on demand, not by default.

The fixture itself is on S3, in the 'test-pfb-inputs' bucket, and the script
automatically downloads it.

Also in S3: the supporting files, like downloadable geometries and GeoJSON
files of destination locations, that get saved to S3 and used by the front end
in various ways. Those have also been added to the 'test-pfb-inputs' bucket
and 'update' has an `aws s3 sync` command to copy them to the user's
development S3 bucket.