Also sync neighborhood boundary files for analysis fixture
The boundary file for a neighborhood doesn't get used by the front end,
but it does get used when you run a new analysis for that neighborhood.
So they also need to be synced into the dev S3 bucket when setting up a
new dev instance and installing hte fixture.
I've put the files in the 'test-pfb-inputs' bucket, this adds a command
to 'update' to sync them to the dev bucket, along with the results files.