Revise README for removal of Vagrant
Removes all the documentation for how to run the dev environment in Vagrant,
which simplifies things quite a bit because there are no more "if you're doing
it this way do this, otherwise this" considerations, just the one way to do it.

I also removed references to "if you don't have AWS credentials" because it
seems like most things stop working without credentials, and there's no reason
to spend time restoring and maintaining that option.